abstract out displayDocumentation getBuildProperties
